+++ src/ui_critplot/about.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 ################################################################################
-## Form generated from reading UI file 'about.ui'
+## Form generated from reading UI file 'about.ui_critplot'
 ##
 ## Created by: Qt User Interface Compiler version 5.15.2
 ##
