+++ src/ui_critplot/about.ui
