+++ src/ui_critplot/critplotform.ui
