+++ tests/opengl/conftest.py
@@ -1,7 +1,7 @@
 from qtpy.QtCore import QCoreApplication, Qt
 from qtpy.QtWidgets import QApplication
 from qtbased.guiopengl_cp import GuiOpenGLCP
-from qtbased.mainform import MainForm
+from qtbased.critplotform import MainForm
 from src.models.cp_model import AtomicModelCP
 
 import pytest
